+++ tests/mod.rs
@@ -17,45 +17,47 @@ extern crate slog_stdlog;
 extern crate slog_envlogger;
 
 use elastic_responses::*;
-use elastic_responses::error::ApiError;
+use elastic_responses::error::*;
 use std::fs::File;
-use std::io::Read;
+use std::io::{Read, Cursor};
 
-fn load_file(p :&str) -> String {
+fn load_file_as_response(status: u16, p: &str) -> HttpResponse<Cursor<Vec<u8>>> {
     let mut f = File::open(p).unwrap();
-    let mut s = String::new();
-    f.read_to_string(&mut s).unwrap();
-    s
+    let mut s = Vec::new();
+    f.read_to_end(&mut s).unwrap();
+    
+    HttpResponse::new(status, Cursor::new(s))
 }
 
 #[test]
 fn test_parse_hits_simple() {
-    let s = load_file("tests/samples/hits_only.json");
-    let deserialized: SearchResponse = serde_json::from_str(&s).unwrap();
+    let s = load_file_as_response(200, "tests/samples/hits_only.json");
+
+    let deserialized = SearchResponse::from_response(s).unwrap();
 
     assert_eq!(deserialized.hits().into_iter().count(), 5);
 }
 
 #[test]
 fn test_parse_simple_aggs() {
-    let s = load_file("tests/samples/aggregation_simple.json");
-    let deserialized: SearchResponse = serde_json::from_str(&s).unwrap();
+    let s = load_file_as_response(200, "tests/samples/aggregation_simple.json");
+    let deserialized = SearchResponse::from_response(s).unwrap();
 
     assert_eq!(deserialized.aggs().into_iter().count(), 124);
 }
 
 #[test]
 fn test_parse_3level_aggs() {
-    let s = load_file("tests/samples/aggregation_3level.json");
-    let deserialized: SearchResponse = serde_json::from_str(&s).unwrap();
+    let s = load_file_as_response(200, "tests/samples/aggregation_3level.json");
+    let deserialized = SearchResponse::from_response(s).unwrap();
 
     assert_eq!(deserialized.aggs().into_iter().count(), 201);
 }
 
 #[test]
 fn test_parse_3level_multichild_aggs() {
-    let s = load_file("tests/samples/aggregation_3level_multichild.json");
-    let deserialized: SearchResponse = serde_json::from_str(&s).unwrap();
+    let s = load_file_as_response(200, "tests/samples/aggregation_3level_multichild.json");
+    let deserialized = SearchResponse::from_response(s).unwrap();
 
     let min = "min_ack_pkts_sent";
     let avg = "avg_ack_pkts_sent";
@@ -76,8 +78,8 @@ fn test_parse_3level_multichild_aggs() {
 
 #[test]
 fn test_parse_3level_multistats_aggs() {
-    let s = load_file("tests/samples/aggregation_3level_multistats.json");
-    let deserialized: SearchResponse = serde_json::from_str(&s).unwrap();
+    let s = load_file_as_response(200, "tests/samples/aggregation_3level_multistats.json");
+    let deserialized = SearchResponse::from_response(s).unwrap();
 
     let min = "extstats_ack_pkts_sent_min";
     let avg = "stats_ack_pkts_sent_avg";
@@ -100,8 +102,8 @@ fn test_parse_3level_multistats_aggs() {
 
 #[test]
 fn test_parse_simple_aggs_no_empty_first_record() {
-    let s = load_file("tests/samples/aggregation_simple.json");
-    let deserialized: SearchResponse = serde_json::from_str(&s).unwrap();
+    let s = load_file_as_response(200, "tests/samples/aggregation_simple.json");
+    let deserialized = SearchResponse::from_response(s).unwrap();
 
     let s = "timechart";
     let mut first = true;
@@ -115,8 +117,8 @@ fn test_parse_simple_aggs_no_empty_first_record() {
 
 #[test]
 fn test_parse_found_doc_response() {
-    let s = load_file("tests/samples/get_found.json");
-    let deserialized: GetResponse = serde_json::from_str(&s).unwrap();
+    let s = load_file_as_response(200, "tests/samples/get_found.json");
+    let deserialized = GetResponse::from_response(s).unwrap();
 
     let id = deserialized.source
         .unwrap()
@@ -133,19 +135,19 @@ fn test_parse_found_doc_response() {
 
 #[test]
 fn test_parse_not_found_doc_response() {
-    let s = load_file("tests/samples/get_not_found.json");
-    let deserialized: GetResponse = serde_json::from_str(&s).unwrap();
+    let s = load_file_as_response(404, "tests/samples/get_not_found.json");
+    let deserialized = GetResponse::from_response(s).unwrap();
 
     assert!(!deserialized.found);
 }
 
 #[test]
 fn test_parse_index_not_found_error() {
-    let s = load_file("tests/samples/error_index_not_found.json");
-    let deserialized: ApiError = serde_json::from_str(&s).unwrap();
+    let s = load_file_as_response(404, "tests/samples/error_index_not_found.json");
+    let deserialized = GetResponse::from_response(s).unwrap_err();
 
     let valid = match deserialized {
-        ApiError::IndexNotFound { ref index }
+        ResponseError::Api(ApiError::IndexNotFound { ref index })
         if index == "carrots" => true,
         _ => false
     };
@@ -155,11 +157,11 @@ fn test_parse_index_not_found_error() {
 
 #[test]
 fn test_parse_parsing_error() {
-    let s = load_file("tests/samples/error_parsing.json");
-    let deserialized: ApiError = serde_json::from_str(&s).unwrap();
+    let s = load_file_as_response(400, "tests/samples/error_parsing.json");
+    let deserialized = SearchResponse::from_response(s).unwrap_err();
 
     let valid = match deserialized {
-        ApiError::Parsing { line: 2, col: 9, ref reason } 
+        ResponseError::Api(ApiError::Parsing { line: 2, col: 9, ref reason })
         if reason == "Unknown key for a START_OBJECT in [qry]." => true,
         _ => false
     };
@@ -169,11 +171,11 @@ fn test_parse_parsing_error() {
 
 #[test]
 fn test_parse_other_error() {
-    let s = load_file("tests/samples/error_other.json");
-    let deserialized: ApiError = serde_json::from_str(&s).unwrap();
+    let s = load_file_as_response(500, "tests/samples/error_other.json");
+    let deserialized = SearchResponse::from_response(s).unwrap_err();
 
     let reason = match deserialized {
-        ApiError::Other(ref err) => err.as_object()
+        ResponseError::Api(ApiError::Other(ref err)) => err.as_object()
                                        .and_then(|err| err.get("reason"))
                                        .and_then(|reason| reason.as_str())
                                        .map(|reason| reason.to_owned()),
